cards folder contains a sample card image for each of 52 cards.
BlackjackComparer.py utilizes this info to create histograms of keypoint distributions which can be compared to new card images to guess what value the card is.
This process is working, but flawed.

BlackjackCard has a name attribute which can be determined by the comparer. The name is displayed in the output